// app/(modules)/menulog/fctrllog/fctrllog-types.ts
export interface Controllist {
    indaprcompmixtas: string;
    indcatobras:      string;
    indcontratacion:  string;
    indestreserva:    string;
    indpac:           string;
    indvalalmdest:    string;
    indvalccostoppto: string;
    raw_rnum_:        number;
    sistcontratacion: string;
   }